requirements.txt
PyYAML==6.0
accelerate @ git+https://github.com/huggingface/accelerate.git@7b4d12623a1847bf7f749d6824017683922339eb
datasets==2.12.0
deepspeed @ git+https://github.com/microsoft/DeepSpeed.git@3f5e4931098bf533f8217afb6d986c90f81aed80
einops==0.6.1
evaluate==0.4.0
peft @ git+https://github.com/huggingface/peft.git@0b62b4378b4ce9367932c73540349da9a41bdea8
protobuf==4.23.3
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
torch==2.0.1
transformers @ git+https://github.com/huggingface/transformers@de9255de27abfcae4a1f816b904915f0b1e23cd9
fire==0.5.0
mpi4py==3.1.4
pandas==2.0.3
Pillow==10.0.0
